+++ level-editor/src/UI/ElementHierarchy.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace Tabby {
+
+class ElementHierarchy {
+public:
+    ElementHierarchy();
+};
+
+}