+++ tabby/src/Tabby/World/Map/MapRegistry.cpp
@@ -1,0 +1,13 @@
+#include <Tabby/World/Map/MapRegistry.h>
+
+namespace Tabby {
+
+MapRegistry::MapRegistry()
+{
+}
+
+MapRegistry::MapRegistry(const std::string& path)
+{
+}
+
+}